backend/requirements.txt
fastapi
uvicorn
python-multipart
numpy
opencv-python
cloudinary
python-dotenv
ultralytics
pymongo
